mem_set_read_only
There is a bug where mem_set_read_only is unable to set a page as read_only. This is tested in kernel/ox_main.c at line 138:

  132    *addr = '1';
  133    printk("addr [%c]\n",*addr);
  134    // Test page allocation and setting of read-only pages.
  135    // Try setting read-only on kernel page.
  136    addr = page_alloc(1);
  137    printk(" ADDR = [%d]\n",addr);
  138    mem_set_read_only(addr,1);
  139    printk("Should crash here\n");
  140    *addr = '2';
  141    printk("Didn't crash\n");
  142    printk("addr [%c]\n",*addr);
  143    *addr = 0;
  144    printk("Didn't crash\n");

The code in Roger master branch keeps going after line 140. Applying this patch the machine stops upon execution of line 140. I'm not sure if it stops for the right reason which I think would be a page fault.

Which would be the right debug technique to use here? Since the machine stops it seems kind of hard to know why ?

Thank you for your comments.